更新 HookTool

+++ app/src/main/java/com/hchen/appretention/hook/hyper/HyperV1.java
@@ -296,7 +296,7 @@ public class HyperV1 extends BaseHC {
                     @Override
                     public void before() {
                         Object config = getArgs(0);
-                        int mPolicy = callMethod(config, getPolicy);
+                        int mPolicy = (int) callMethod(config, getPolicy);
                         if (!ProcessPolicy.getKillReason(mPolicy).equals(ProcessPolicy.REASON_GARBAGE_CLEAN))
                             returnNull();
                     }
